Organized repo and updated docs
