+++ build-windows.bat
@@ -1,56 +1,68 @@
 @echo off
-setlocal enabledelayedexpansion
+setlocal EnableDelayedExpansion
 
 :: Check if required arguments are provided
 if "%~1"=="" (
     echo Error: Missing CUDA Toolkit path
-    echo Usage: %0 ^<CUDA_Toolkit_Path^> ^<CMAKE_PREFIX_PATH^>
+    echo Usage: .\build-windows.bat CUDA_PATH CMAKE_PREFIX_PATH [delete]
     exit /b 1
 )
 
 if "%~2"=="" (
     echo Error: Missing CMAKE_PREFIX_PATH
-    echo Usage: %0 ^<CUDA_Toolkit_Path^> ^<CMAKE_PREFIX_PATH^>
+    echo Usage: .\build-windows.bat CUDA_PATH CMAKE_PREFIX_PATH [delete]
     exit /b 1
 )
 
 set CUDA_PATH=%~1
 set CMAKE_PREFIX_PATH=%~2
+set DELETE_BUILD=0
 
-:: Delete the build directory if it exists
+:: Check if third parameter is 'delete'
+if /i "%~3"=="delete" (
+    set DELETE_BUILD=1
+)
+
+:: Build directory handling
 if exist build (
-    echo Removing existing build directory...
-    rd /s /q build
-    if errorlevel 1 (
-        echo Failed to remove build directory. Exiting.
+    if !DELETE_BUILD!==1 (
+        echo Removing existing build directory...
+        rd /s /q build
+        if errorlevel 1 (
+            echo Failed to remove build directory. Exiting.
+            exit /b 1
+        )
+        echo Build directory removed.
+    ) else (
+        echo ERROR: Build directory exists! Add 'delete' parameter to remove it.
+        echo Example: .\build-windows.bat CUDA_PATH CMAKE_PREFIX_PATH delete
         exit /b 1
     )
-    echo Build directory removed successfully.
 ) else (
-    echo No existing build directory found.
+    echo No existing build directory.
 )
 
 :: Create build directory
-echo Creating new build directory...
+echo Creating build directory...
 mkdir build
 if errorlevel 1 (
-    echo Failed to create build directory. Exiting.
+    echo Failed to create build directory.
     exit /b 1
 )
 
-:: Run CMake configure with provided arguments
-echo Configuring project with CMake...
+:: Run CMake configure
+echo Running CMake configuration...
 cmake --preset x64-release -DCUDAToolkit_ROOT="%CUDA_PATH%" -DCMAKE_PREFIX_PATH="%CMAKE_PREFIX_PATH%"
 if errorlevel 1 (
-    echo CMake configuration failed. Exiting.
+    echo CMake configuration failed.
     exit /b 1
 )
 
-:: Change to the build directory
+:: Change to build directory
 echo Changing to build directory...
 cd .\build\x64-release\
 if errorlevel 1 (
-    echo Failed to change to build directory. Exiting.
+    echo Failed to change directory.
     exit /b 1
 )
